refactor(tui): group semantic scroll modules
