+++ station/src/input/stationInput.ts
@@ -13,7 +13,7 @@ import { sanitizePastedText } from "../station/input/sequenceToTuiKey.js";
 import { dispatchStationKey } from "../station/input/stationActions.js";
 import type { StationClientStateSource } from "@station/client";
 import type { DashboardActions, DashboardStateSource } from "@station/dashboard-core/runtime";
-import type { DashboardScrollController } from "../station/view/layout/dashboardScrollController.js";
+import type { DashboardScrollController } from "../station/view/layout/scroll/dashboardScrollController.js";
 import {
   routeKey,
   routeMouse,
