+++ station/src/station/view/settings/ProjectSettingsPanelView.tsx
@@ -18,7 +18,7 @@ import type {
 } from "@station/dashboard-core/state";
 import { toOpenTuiColor, useStationTheme } from "../../../theme/index.js";
 import { EditableTextInputView } from "../EditableTextInputView.js";
-import { semanticItemRenderableId } from "../layout/scrollViewport.js";
+import { semanticItemRenderableId } from "../layout/scroll/scrollViewport.js";
 import { AgentChoiceListView } from "../sheets/AgentChoiceListView.js";
 import { fit } from "../sheets/parts.js";
 import {
