+++ station/src/station/view/settings/SettingsPanelView.tsx
@@ -3,8 +3,8 @@ import type { SettingsPanelFocus } from "@station/dashboard-core/state";
 import type { ReactNode } from "react";
 import { toOpenTuiColor, toOpenTuiOpaqueColor, useStationTheme } from "../../../theme/index.js";
 import type { StationMouseTarget } from "../../input/stationMouse.js";
-import { SemanticScrollRegion } from "../layout/SemanticScrollViewport.js";
-import { semanticItemRenderableId } from "../layout/scrollViewport.js";
+import { SemanticScrollRegion } from "../layout/scroll/SemanticScrollViewport.js";
+import { semanticItemRenderableId } from "../layout/scroll/scrollViewport.js";
 import { settingsPanelFrame } from "../layout/settingsPanelFrame.js";
 import { fit } from "../sheets/parts.js";
 import {
